Add a keyModel option and bump version
The keyModel option can be added as an attribute to the directive:
<cp=select ng-model="model" collection="collection" key-model></cp-select>

This will expect the model value to be the "key" attribute of the collection of model options.
For example:
var collection = [
  {
		key: 'UT',
		value: 'Utah'
	}
]

If the key-model attribute was set, then when selecting the option out of the collection "UT" would be
the selected model.